🔥 Revert to default behaviour of Nexmo
Code was added to use Rails Credentials for Nexmo, but from a library point of view, this should not come from Rails.

Nexmo uses env vars by default. Env vars should be set from the Rails Credentials instead, and then Nexmo will be initialized using these env vars.

This is needed for keeping compatibility with other users of the library (for instance Amaretti).